+++ src/Commands/SystemCommands/EditedchannelpostCommand.php
@@ -1,0 +1,51 @@
+<?php
+/**
+ * This file is part of the TelegramBot package.
+ *
+ * (c) Avtandil Kikabidze aka LONGMAN <akalongman@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Longman\TelegramBot\Commands\SystemCommands;
+
+use Longman\TelegramBot\Commands\SystemCommand;
+
+/**
+ * Edited channel post command
+ *
+ * @todo Remove due to deprecation!
+ */
+class EditedchannelpostCommand extends SystemCommand
+{
+    /**
+     * @var string
+     */
+    protected $name = 'editedchannelpost';
+
+    /**
+     * @var string
+     */
+    protected $description = 'Handle edited channel post';
+
+    /**
+     * @var string
+     */
+    protected $version = '1.0.0';
+
+    /**
+     * Execute command
+     *
+     * @return mixed
+     * @throws \Longman\TelegramBot\Exception\TelegramException
+     */
+    public function execute()
+    {
+        //$edited_channel_post = $this->getUpdate()->getEditedChannelPost();
+
+        trigger_error(__CLASS__ . ' is deprecated and will be removed and handled by ' . GenericmessageCommand::class . ' by default in a future release.', E_USER_DEPRECATED);
+
+        return parent::execute();
+    }
+}